refactor(book-ui): rename to .jsx, simplify Navbar and Vite config
- Rename 29 JSX-containing .js files to .jsx
- Simplify vite.config.js by removing esbuild/optimizeDeps workarounds
- Rewrite Navbar removing burger menu for a simpler implementation
- Remove unused  from 28 files
