+++ README.md
@@ -52,7 +52,7 @@ On [ivangfr.github.io](https://ivangfr.github.io), I have compiled my Proof-of-C
   
   To log in, a `user` or `admin` must provide valid `username` and `password` credentials. `book-ui` communicates with `book-api` to get `books` and `users` data.
   
-  `book-ui` uses [`Mantine`](https://mantine.dev/) as CSS-styled framework.
+  `book-ui` uses [`Mantine`](https://mantine.dev/) as a UI component library, with [`Tabler Icons`](https://tabler.io/icons) for icons.
 
 ## Prerequisites
