+++ book-ui/index.html
@@ -10,6 +10,6 @@
   </head>
   <body>
     <div id="root"></div>
-    <script type="module" src="/src/index.js"></script>
+    <script type="module" src="/src/index.jsx"></script>
   </body>
 </html>
