+++ book-ui/src/components/context/AuthContext.test.jsx
@@ -1,4 +1,3 @@
-import React from 'react'
 import { render, screen, act } from '@testing-library/react'
 import { AuthProvider, useAuth } from './AuthContext'
