+++ book-ui/src/components/home/Home.test.jsx
@@ -1,4 +1,3 @@
-import React from 'react'
 import { screen, waitFor } from '@testing-library/react'
 import { render } from '../../test-utils'
 import Home from './Home'
