+++ book-ui/src/components/home/Signup.jsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react'
+import { useState } from 'react'
 import { NavLink, Navigate } from 'react-router-dom'
 import { TextInput, PasswordInput, Button, Paper, Title, Text, Center, Stack, Alert } from '@mantine/core'
 import { IconAlertCircle } from '@tabler/icons-react'
