+++ book-ui/src/components/misc/BookCover.jsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react'
+import { useState } from 'react'
 import { Image, Skeleton } from '@mantine/core'
 
 function BookCover({ isbn, w, h }) {
