+++ book-ui/src/components/misc/PrivateRoute.jsx
@@ -1,4 +1,3 @@
-import React from 'react'
 import { Navigate } from 'react-router-dom'
 import { useAuth } from '../context/AuthContext'
