+++ book-ui/src/components/misc/PrivateRoute.test.jsx
@@ -1,4 +1,3 @@
-import React from 'react'
 import { screen } from '@testing-library/react'
 import { render } from '../../test-utils'
 import PrivateRoute from './PrivateRoute'
