+++ book-ui/src/components/user/BookList.test.jsx
@@ -1,4 +1,3 @@
-import React from 'react'
 import { screen } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
 import { render } from '../../test-utils'
