+++ book-ui/src/test-utils.jsx
@@ -1,4 +1,3 @@
-import React from 'react'
 import { render } from '@testing-library/react'
 import { MantineProvider } from '@mantine/core'
 import { MemoryRouter } from 'react-router-dom'
